+++ lib/main.dart
@@ -14,7 +14,6 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:MobileOne/pages/authentication-page.dart';
 import 'package:MobileOne/localization/delegate.dart';
-import 'package:MobileOne/localization/localization.dart';
 import 'package:MobileOne/localization/supported.dart';
 import 'package:MobileOne/pages/Mainpage.dart';
 import 'package:MobileOne/pages/create_list.dart';
@@ -30,12 +29,7 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:get_it/get_it.dart';
 import 'package:google_sign_in/google_sign_in.dart';
-import 'package:http/http.dart' as http;
-import 'package:provider/provider.dart';
-import 'package:webfeed/webfeed.dart';
 import 'package:firebase_crashlytics/firebase_crashlytics.dart';
-
-import 'localization/localization.dart';
 
 GetIt getIt = GetIt.instance;
 
@@ -109,69 +103,6 @@ class MyApp extends StatelessWidget {
         "/cards": (contaxt) => Cards(),
         "/loyaltycards": (contaxt) => LoyaltyCards(),
       },
-    );
-  }
-}
-
-class HelloPage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Text(getString(context, 'app_name')),
-            RaisedButton(
-              onPressed: () => openMainPage(context),
-              child: Text(getString(context, 'open_main_page')),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  void openMainPage(context) {
-    Navigator.of(context).pushNamed(
-      '/mainpage',
-    );
-  }
-}
-
-class RssProvider extends ChangeNotifier {
-  RssFeed _rssFeed = RssFeed();
-  RssFeed get feed => _rssFeed;
-}
-
-Future<RssFeed> fetchRssFeed() async {
-  var response = await http.get('https://itsallwidgets.com/podcast/feed');
-  print('Response status: ${response.statusCode}');
-  print('Response body: ${response.body}');
-  return new RssFeed.parse(response.body);
-}
-
-class ProviderPage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: FutureProvider(
-        create: (_) => fetchRssFeed(),
-        child: Consumer(
-          builder: (context, RssFeed rss, _) {
-            if (rss == null) {
-              return Center(child: CircularProgressIndicator());
-            }
-            rss.items.sort((a, b) => a.title.compareTo(b.title));
-            return ListView.builder(
-              itemBuilder: (context, index) {
-                return Text(rss.items[index].title);
-              },
-              itemCount: rss.items.length,
-            );
-          },
-        ),
-      ),
     );
   }
 }